refactor: extract settings page widgets
