+++ lib/widgets/settings_segments/theme_dropdown.dart
@@ -1,0 +1,55 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:pomo/l10n/l10n.dart';
+import 'package:pomo/pages/settings/cubit/settings_cubit.dart';
+
+class ThemeDropdown extends StatelessWidget {
+  const ThemeDropdown({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final l10n = context.l10n;
+
+    return BlocBuilder<SettingsCubit, SettingsState>(
+      builder: (context, state) {
+        return Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 16),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            children: [
+              Text(
+                l10n.themeMode,
+                style: Theme.of(context).textTheme.labelLarge,
+              ),
+              const SizedBox(height: 8),
+              DropdownButton<ThemeMode>(
+                value: state.themeMode,
+                isExpanded: true,
+                items: [
+                  DropdownMenuItem<ThemeMode>(
+                    value: ThemeMode.light,
+                    child: Text(l10n.light),
+                  ),
+                  DropdownMenuItem<ThemeMode>(
+                    value: ThemeMode.dark,
+                    child: Text(l10n.dark),
+                  ),
+                  DropdownMenuItem<ThemeMode>(
+                    value: ThemeMode.system,
+                    child: Text(l10n.system),
+                  ),
+                ],
+                onChanged: (val) => context
+                    .read<SettingsCubit>()
+                    .setThemeMode(val ?? ThemeMode.system),
+              ),
+            ],
+          ),
+        );
+      },
+    );
+  }
+}
